script/requirements.txt
nats.py==2.2.0